requirements.txt
yfinance==0.2.36
tavily-python==0.3.1
openai==1.12.0
streamlit==1.33.0
plotly==5.18.0
python-dotenv==1.0.1